requirements.txt
# Voice-to-Task Meeting Assistant
# Python 3.11+ Required

# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
websockets==12.0
python-socketio==5.11.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.2.5

# Redis & Celery
celery[redis]==5.3.6
redis==5.0.1
flower==2.0.1

# AI & LLM
openai==1.12.0
langchain==0.1.9
langchain-openai==0.0.6
langchain-core==0.1.27

# Audio Processing
openai-whisper==20231117
pydub==0.25.1
soundfile==0.12.1
librosa==0.10.1
pyannote.audio==3.1.1

# HTTP & External APIs
httpx==0.26.0
aiohttp==3.9.3
requests==2.31.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation & Serialization
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Utilities
numpy==1.26.4
pandas==2.2.0
python-dateutil==2.8.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0

# Logging & Monitoring
structlog==24.1.0
sentry-sdk[fastapi]==1.40.0

# Development
black==24.1.1
ruff==0.2.1
mypy==1.8.0
pre-commit==3.6.0
